% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MatchSummaries}
\alias{MatchSummaries}
\title{UEFA Euro 2020 Match summaries}
\format{
A data frame with 51 rows and 15 variables
\describe{
  \item{matchday}{numeric from 1 (1st group phase match) to 7 (final)}
  \item{date}{date of match}
  \item{home}{home team name}
  \item{away}{away team name}
  \item{sc_home}{number of goals after regular or extended time for the home team}
  \item{sc_away}{number of goals after regular or extended time for the away team}
  \item{had_penalties}{logical indicator for penalty shootout}
  \item{sc_pen_home}{number of score penalties for the home team}
  \item{sc_pen_away}{number of score penalties for the away team}
  \item{pos_home}{ball possession for the home team}
  \item{pos_away}{ball possession for the home team}
  \item{total_shots_home}{number of attempted shots for the home team}
  \item{total_shots_away}{number of attempted shots for the away team}
  \item{shots_on_target_home}{number of shots on target for the home team}
  \item{shots_on_target_away}{number of shots on target for the away team}
}
}
\usage{
MatchSummaries
}
\description{
Match-level data for 51 soccer matches played during UEFA Euro 2020
}
\keyword{datasets}
