% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passes_by_match.R
\name{passes_by_match}
\alias{passes_by_match}
\title{Obtain passing networks for both teams by match}
\usage{
passes_by_match(home_team, away_team)
}
\arguments{
\item{home_team}{Name of home team}

\item{away_team}{Name of away team}
}
\value{
Named list of two igraph objects.
}
\description{
Obtain passing networks for both teams by match
}
\examples{
\dontrun{
passes_by_match('Italy', 'England')
}
}
