% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passes_by_team.R
\name{passes_by_team}
\alias{passes_by_team}
\title{Obtain passing networks for all the matches of a team}
\usage{
passes_by_team(team)
}
\arguments{
\item{team}{Name of team to return passing networks for}
}
\value{
Named list of n igraph objects sorted by matchday.
}
\description{
Obtain passing networks for all the matches of a team
}
\examples{
\dontrun{
passes_by_team('Italy')
}
}
