% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PassNetworks}
\alias{PassNetworks}
\title{UEFA Euro 2020 Pass networks}
\format{
A data frame with 51 rows and 4 variables
\describe{
  \item{home}{home team name}
  \item{away}{away team name}
  \item{home_graph}{
  list column with \code{igraph} objects of directed pass networks for the home teams

  \bold{Attributes}
    \describe{
       \item{name}{Player name}
       \item{mins_played}{Minutes played until/after substitution}
       \item{pc_long}{Number of completed long passes}
       \item{pa_long}{Number of attempted long passes}
       \item{pc_medium}{Number of completed medium passes}
       \item{pa_medium}{Number of attempted medium passes}
       \item{pc_short}{Number of completed short passes}
       \item{pa_short}{Number of attempted short passes}
       \item{pc_total}{Number of completed passes}
       \item{pa_total}{Number of attempted passes}
       \item{pass_accuracy_total}{Completed passes / attempted passes}
    }
  }
  \item{away_graph}{equivalent of home_graph for away team}
}
}
\usage{
PassNetworks
}
\description{
Match-level passing distributions for 51 soccer matches played during UEFA Euro 2020.
}
\keyword{datasets}
